+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Swarm copy POC.
+- Agent memory.
 
 
 ## [0.2.0] - 22.10.2024